refactor(graph): rename lib/falkordb -> lib/people-graph; remove unused FalkorDB deps
Pure rename now that nothing FalkorDB remains:
- git mv src/lib/falkordb -> src/lib/people-graph (career-signals, name-matching,
  people, scoring, suggestions, telemetry) and update all import paths
- rename tests/falkordb-*.test.ts -> tests/{people-graph,career-signals}.test.ts
- resetFalkorTelemetryForTests -> resetSuggestionTelemetryForTests
- drop unused 'falkordb' + 'falkordblite' npm deps (package.json + serverExternalPackages)
  and update lockfile; the only importer was the deleted client.ts
- collapse the vestigial 'falkor' suggestion mode/telemetry (mode is always
  sql_fallback now; removed falkorCount) and de-Falkor stray comments

No behavior change. typecheck + lint + full fast suite (6025 tests) green.
