+++ apps/web/src/lib/mentorship/signal-backfill.ts
@@ -7,7 +7,7 @@ import {
   canonicalizeIndustry,
   canonicalizeRoleFamily,
   normalizeCareerText,
-} from "@/lib/falkordb/career-signals";
+} from "@/lib/people-graph/career-signals";
 import { extractSignalsFromGoals } from "@/lib/mentorship/goals-extraction";
 import { createHash } from "crypto";
